requirements.txt
anyio==3.7.1
appdirs==1.4.4
beautifulsoup4==4.12.2
Bottleneck
certifi==2023.7.22
charset-normalizer==3.2.0
click
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.2
fonttools==4.42.0
frozendict==2.3.8
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.4
joblib
kiwisolver==1.4.4
lxml==4.9.3
matplotlib==3.7.2
multitasking==0.0.11
numexpr
numpy==1.25.2
packaging==23.1
pandas
plotly
polygon==1.1.0
python-dateutil
python-dotenv
pytz
requests==2.31.0
six
sniffio==1.3.0
soupsieve==2.5
tenacity
tzdata
urllib3==2.0.4
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
yfinance==0.2.28
